Re-estructurando el repo
