+++ system/src/main/java/com/mujeres2000/system/Configuration/SwaggerConfig.java
@@ -1,0 +1,8 @@
+package com.mujeres2000.system.Configuration;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+
+public class SwaggerConfig {
+}
