+++ system/src/main/java/com/mujeres2000/system/Exception/BadRequestException.java
@@ -1,0 +1,13 @@
+package com.mujeres2000.system.Exception;
+
+public class BadRequestException  extends RuntimeException{
+    //Manejo de excepciones
+    //CONTRUCTORES
+
+    public BadRequestException() {
+    }
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
