+++ system/src/main/java/com/mujeres2000/system/Exception/NotFoundException.java
@@ -1,0 +1,13 @@
+package com.mujeres2000.system.Exception;
+
+public class NotFoundException extends RuntimeException{
+    //Manejo de excepciones
+    //CONTRUCTORES
+
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
