+++ system/src/main/java/com/mujeres2000/system/Exception/UnauthorizedException.java
@@ -1,0 +1,13 @@
+package com.mujeres2000.system.Exception;
+
+public class UnauthorizedException extends RuntimeException {
+    //Manejo de excepciones
+    //CONTRUCTORES
+
+    public UnauthorizedException() {
+    }
+
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
