+++ system/src/main/java/com/mujeres2000/system/Repository/ProductoRepository.java
@@ -1,0 +1,9 @@
+package com.mujeres2000.system.Repository;
+
+import com.mujeres2000.system.model.Producto;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductoRepository extends JpaRepository<Producto,Integer> {
+}
