--- system/system/src/main/java/com/mujeres2000/system/Repository/CostoGeneralRepository.java
@@ -1,8 +1,0 @@
-package com.mujeres2000.system.Repository;
-
-
-import com.mujeres2000.system.model.CostoGeneral;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CostoGeneralRepository extends JpaRepository<CostoGeneral,Integer> {
-}
