--- system/system/src/test/java/com/mujeres2000/system/Mujeres2000SystemApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mujeres2000.system;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Mujeres2000SystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
